Refactor: split utility class
