+++ src/test/java/org/cdlib/util/marc2/MarcHelperTest.java
@@ -4,7 +4,6 @@ import static org.junit.Assert.*;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.function.Predicate;
 import org.cdlib.util.DeserializationException;
 import org.cdlib.util.FileUtil;
 import org.cdlib.util.marc2.MarcRecordHelper;
@@ -72,81 +71,6 @@ public class MarcHelperTest {
   public void controlFieldVal_happyPath() {
     String expected = "751666086";
     assertEquals(expected, marcHelper.controlFieldVal("001").get());
-  }
-
-  @Test
-  public void marcDataFields_happyPath() {
-    Optional<List<MarcDataField>> result = marcHelper.marcDataFields("020");
-    List<MarcDataField> resultList = result.get();
-    assertEquals(2, resultList.size());
-    assertEquals("9781847171481", resultList.get(0)
-                                            .getSubFields()
-                                            .get('a')
-                                            .get()
-                                            .get(0));
-  }
-
-  @Test
-  public void marcDataFieldsWithFilter_happyPath() {
-    Predicate<MarcDataField> test = (df) -> df.getSubFields()
-                                              .get('a')
-                                              .get()
-                                              .get(0)
-                                              .equals("9781847171481");
-    Optional<List<MarcDataField>> result = marcHelper.marcDataFields("020", test);
-    List<MarcDataField> resultList = result.get();
-    assertEquals(1, resultList.size());
-    assertEquals("9781847171481", resultList.get(0)
-                                            .getSubFields()
-                                            .get('a')
-                                            .get()
-                                            .get(0));
-  }
-
-  @Test
-  public void marcDataFieldsWithFilterSubfieldNotFound_emptyResult() {
-    Optional<List<MarcDataField>> result = marcHelper.marcDataFields("020", hasSubfieldMatching('x', "9781847171481"));
-    List<MarcDataField> resultList = result.get();
-    assertEquals(0, resultList.size());
-  }
-  
-  @Test
-  public void marcDataFieldsWithControlFieldTag_emptyResult() {
-    Optional<List<MarcDataField>> result = marcHelper.marcDataFields("008", hasSubfieldMatching('x', "9781847171481"));
-    List<MarcDataField> resultList = result.get();
-    assertEquals(0, resultList.size());
-  }
-  
-  @Test
-  public void marcDataFieldsWithFilterTagNotFound_emptyResult() {
-    Optional<List<MarcDataField>> result = marcHelper.marcDataFields("999", hasSubfieldMatching('a', "9781847171481"));
-    List<MarcDataField> resultList = result.get();
-    assertEquals(0, resultList.size());
-  }
-  
-  @Test
-  public void marcDataFieldsWithFilterTagNull_emptyResult() {
-    Optional<List<MarcDataField>> result = marcHelper.marcDataFields(null, hasSubfieldMatching('a', "9781847171481"));
-    List<MarcDataField> resultList = result.get();
-    assertEquals(0, resultList.size());
-  }
-
-  private Predicate<MarcDataField> hasSubfieldMatching(char key, String val) {
-    return (df) -> {
-      Optional<List<String>> resultVal = df.getSubFields().get(key);
-      return resultVal.map(list -> list.get(0)
-                                       .equals(val))
-                      .orElse(false);
-    };
-
-  }
-
-  @Test
-  public void marcDataFieldsNotFound_emptyList() {
-    Optional<List<MarcDataField>> result = marcHelper.marcDataFields("776");
-    List<MarcDataField> resultList = result.get();
-    assertTrue(resultList.isEmpty());
-
   }
 
   @Test(expected = IllegalArgumentException.class)
